agent_app/pyproject.toml
[project]
name = "agent-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask[async]>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain[google-genai,openai]>=1.2.10",
    "python-dotenv>=1.2.1",
    "google-generativeai>=0.8.6",
    "arcade-mcp-server>=1.17.0",
    "langchain-google-vertexai>=3.2.2",
]
